+++ EasyPost.Integration/Basics.cs
@@ -34,6 +34,7 @@ public class Basics
         var endShipper = new EndShipper();
         var endShipperCollection = new EndShipperCollection();
         var error = new Error();
+        var estimateDeliveryDateForZipPairResult = new EstimateDeliveryDateForZipPairResult();
         var @event = new Event();
         var eventCollection = new EventCollection();
         var fee = new Fee();
@@ -55,6 +56,8 @@ public class Basics
         var predefinedPackage = new PredefinedPackage();
         var rate = new Rate();
         var rateWithEstimatedDeliveryDate = new RateWithEstimatedDeliveryDate();
+        var recommendShipDateForZipPairResult = new RecommendShipDateForZipPairResult();
+        var recommendShipDateForShipmentResult = new RecommendShipDateForShipmentResult();
         var referralCustomer = new ReferralCustomer();
         var refund = new Refund();
         var report = new Report();
@@ -70,6 +73,9 @@ public class Basics
         var supportedFeature = new SupportedFeature();
         var taxIdentifier = new TaxIdentifier();
         var timeInTransit = new TimeInTransit();
+        var timeInTransitDetails = new TimeInTransitDetails();
+        var timeInTransitDetailsForDeliveryDateEstimate = new TimeInTransitDetailsForDeliveryDateEstimate();
+        var timeInTransitDetailsByDeliveryDate = new TimeInTransitDetailsForShipDateRecommendation();
         var tracker = new Tracker();
         var trackerCollection = new TrackerCollection();
         var trackingDetail = new TrackingDetail();
@@ -139,6 +145,8 @@ public class Basics
         var shipmentInsureParameters = new EasyPost.Parameters.Shipment.Insure();
         var shipmentRegenerateRatesParameters = new EasyPost.Parameters.Shipment.RegenerateRates();
         var shipmentRetrieveEstimatedDeliveryDateParameters = new EasyPost.Parameters.Shipment.RetrieveEstimatedDeliveryDate();
+        var smartRateEstimateDeliveryDateForZipPairParameters = new EasyPost.Parameters.SmartRate.EstimateDeliveryDateForZipPair();
+        var smartRateRecommendShipDateForZipPairParameters = new EasyPost.Parameters.SmartRate.RecommendShipDateForZipPair();
         var taxIdentifierCreateParameters = new EasyPost.Parameters.TaxIdentifier.Create();
         var trackerCreateParameters = new EasyPost.Parameters.Tracker.Create();
         var trackerAllParameters = new EasyPost.Parameters.Tracker.All();