+++ EasyPost/Models/API/RateWithEstimatedDeliveryDate.cs
@@ -11,44 +11,23 @@ namespace EasyPost.Models.API
         #region JSON Properties
 
         /// <summary>
-        ///     The <see cref="Models.API.SmartRate"/> object.
+        ///     The <see cref="Models.API.Rate"/> object.
         /// </summary>
         [JsonProperty("rate")]
         public Rate? Rate { get; set; }
 
         /// <summary>
         ///     Estimated <see cref="TimeInTransitDetails"/> for the <see cref="Rate"/>.
+        ///     Deprecated: Use <see cref="TimeInTransitDetails"/> instead.
+        /// </summary>
+        [Obsolete("This property will be removed in a future version and replaced with TimeInTransitDetails.")]
+        public TimeInTransitDetails? EasyPostTimeInTransitData => this.TimeInTransitDetails?.AsDeprecatedTimeInTransitDetails();
+
+        /// <summary>
+        ///     Estimated <see cref="TimeInTransitDetailsForDeliveryDateEstimate"/> for the <see cref="Rate"/>.
         /// </summary>
         [JsonProperty("easypost_time_in_transit_data")]
-        public TimeInTransitDetails? EasyPostTimeInTransitData { get; set; }
-
-        #endregion
-    }
-
-    /// <summary>
-    ///     Class representing estimated transit times for a <see cref="RateWithEstimatedDeliveryDate"/>.
-    /// </summary>
-    public class TimeInTransitDetails
-    {
-        #region JSON Properties
-
-        /// <summary>
-        ///     Confidence levels for days in transit estimates.
-        /// </summary>
-        [JsonProperty("days_in_transit")]
-        public TimeInTransit? DaysInTransit { get; set; }
-
-        /// <summary>
-        ///     EasyPost's estimated delivery date for the associated <see cref="RateWithEstimatedDeliveryDate"/>.
-        /// </summary>
-        [JsonProperty("easypost_estimated_delivery_date")]
-        public DateTime? EasyPostEstimatedDeliveryDate { get; set; }
-
-        /// <summary>
-        ///     The planned departure date for the shipment.
-        /// </summary>
-        [JsonProperty("planned_ship_date")]
-        public DateTime? PlannedShipDate { get; set; }
+        public TimeInTransitDetailsForDeliveryDateEstimate? TimeInTransitDetails { get; set; }
 
         #endregion
     }